refactor: apply hexagonal architecture
