+++ pom.xml
@@ -113,7 +113,7 @@
                             <generateApiTests>false</generateApiTests>
                             <generateModelTests>false</generateModelTests>
                             <generateSupportingFiles>false</generateSupportingFiles>
-                            <modelPackage>com.fortun.backend.model</modelPackage>
+                            <modelPackage>com.fortun.backend.prices.domain</modelPackage>
                             <typeMappings>
                                 <typeMapping>OffsetDateTime=LocalDateTime</typeMapping>
                             </typeMappings>
@@ -143,7 +143,7 @@
                         <param>com.fortun.backend.model.*</param>
                     </excludedClasses>
                     <excludedMethods>
-                        <param>handler</param>
+                        <param>priceHandler</param>
                         <param>route</param>
                     </excludedMethods>
                 </configuration>
