+++ src/test/java/com/fortun/backend/BackendApplicationTests.java
@@ -1,6 +1,6 @@
 package com.fortun.backend;
 
-import com.fortun.backend.model.PriceResponse;
+import com.fortun.backend.prices.domain.PriceResponse;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
